requirements.txt
certifi==2025.1.31
clickhouse-connect==0.8.15
cx_Oracle==8.3.0
greenlet==3.1.1
lz4==4.4.3
numpy==2.2.3
pandas==2.2.3
psycopg2==2.9.10
PyMySQL==1.1.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zstandard==0.23.0
